feat: added columnar scoring metrics
