+++ crates/abd-clam/src/utils.rs
@@ -3,6 +3,7 @@
 use core::cmp::Ordering;
 
 use distances::{number::Float, Number};
+use rand::prelude::*;
 
 /// Return the number of samples to take from the given population size so as to
 /// achieve linear time complexity for geometric median estimation.
@@ -28,6 +29,18 @@ pub fn num_samples(population_size: usize, sqrt_thresh: usize, log2_thresh: usiz
             }
             .as_usize()
     }
+}
+
+/// Choose a subset of the given items using the given thresholds.
+///
+/// See the `num_samples` function for more information on how the number of
+/// samples is calculated.
+pub fn choose_samples<T: Clone>(indices: &[T], sqrt_thresh: usize, log2_thresh: usize) -> Vec<T> {
+    let mut indices = indices.to_vec();
+    let n = crate::utils::num_samples(indices.len(), sqrt_thresh, log2_thresh);
+    indices.shuffle(&mut rand::thread_rng());
+    indices.truncate(n);
+    indices
 }
 
 /// Returns the number of distinct pairs that can be formed from `n` elements
